Move current user back to global and not from DB.
So, @d4mation brought up that grabbing the current user data from the DB instead of a global fixed some issues with other plugins that falsely modified the current user global. Well, I forgot that I actually modify that global on the Customize preview window, as to show the role being edited. And actually, it does make the most sense to use the global, as it should be editable and then Client Dash should show information based on that. It's unfortunate, but I had to change it back.

Fixes #154