More specific "Tested To" version
Previously we had this set to 5.8 which the WP Plugins Repo
auto-corrects to use the latest Patch version.

While that's super convenient for us, it isn't an accurate reflection of
when it was last tested. And it means that we aren't updating the "Last
Updated" timestamp for the plugin either which is another downside.